Shared dependencies.
The module now uses a specific file (data/package.php) to describe its
dependencies. This file is shared with the stub (which passes them along
to Erebot where they are interpreted by Composer).

This means that the metadata can now be handled from a single file, which
really makes things a lot easier.